requirements.txt
python-telegram-bot>=21.0
openai>=1.0
gspread>=6.0
google-auth>=2.0
python-dotenv>=1.0
